Move classes to other
