+++ sdk/types/scrypted_python/scrypted_sdk/other.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
-from typing import AbstractSet, Any, Callable
+
+from enum import Enum
+from typing import AbstractSet, Any, Callable, Literal
 try:
     from typing import TypedDict
 except:
@@ -7,6 +9,7 @@ except:
 
 SettingValue = str
 EventListener = Callable[[Any, Any, Any], None]
+VibratePattern = int | list[int]
 
 
 class Console:
@@ -36,3 +39,51 @@ class MediaObject:
 
 class RTCSessionDescriptionInit(TypedDict):
     pass
+
+
+class NotificationAction(TypedDict, total=False):
+
+    action: str
+    title: str
+    icon: str  # optional
+
+
+class NotificationDirection(str, Enum):
+
+    auto = "auto"
+    ltr = "ltr"
+    rtl = "rtl"
+
+
+class WebSocket:
+
+    CLOSED: int
+    CLOSING: int
+    CONNECTING: int
+    EventTarget: dict
+    OPEN: int
+    binaryType: Literal["blob", "arraybuffer"]
+    bufferedAmount: int
+    extensions: str
+    onclose: Callable[[dict], None]
+    onerror: Callable[[dict], None]
+    onmessage: Callable[[dict], None]
+    onopen: Callable[[dict], None]
+    protocol: str
+    readyState: int
+    url: str
+
+    def addEventListener(self, type: str, listener: Callable[[dict], None], options: dict = None) -> None:
+        pass
+
+    def close(self, code: int = None, reason: str = None) -> None:
+        pass
+
+    def dispatchEvent(self, event: dict) -> bool:
+        pass
+
+    def removeEventListener(self, type: str, listener: Callable[[dict], None], options: dict = None) -> None:
+        pass
+
+    def send(self, data: str | bytes | bytearray | int | float | bool) -> None:
+        pass
