+++ sdk/types/scrypted_python/scrypted_sdk/types.py
@@ -4,60 +4,9 @@ try:
     from typing import TypedDict
 except:
     from typing_extensions import TypedDict
-from typing import Union, Any, Callable, Literal
+from typing import Union, Any
 
 from .other import *
-
-VibratePattern = int | list[int]
-
-
-class NotificationAction(TypedDict, total=False):
-
-    action: str
-    title: str
-    icon: str  # optional
-
-
-class NotificationDirection(str, Enum):
-
-    auto = "auto"
-    ltr = "ltr"
-    rtl = "rtl"
-
-
-class WebSocket:
-
-    CLOSED: int
-    CLOSING: int
-    CONNECTING: int
-    EventTarget: dict
-    OPEN: int
-    binaryType: Literal["blob", "arraybuffer"]
-    bufferedAmount: int
-    extensions: str
-    onclose: Callable[[dict], None]
-    onerror: Callable[[dict], None]
-    onmessage: Callable[[dict], None]
-    onopen: Callable[[dict], None]
-    protocol: str
-    readyState: int
-    url: str
-
-    def addEventListener(self, type: str, listener: Callable[[dict], None], options: dict = None) -> None:
-        pass
-
-    def close(self, code: int = None, reason: str = None) -> None:
-        pass
-
-    def dispatchEvent(self, event: dict) -> bool:
-        pass
-
-    def removeEventListener(self, type: str, listener: Callable[[dict], None], options: dict = None) -> None:
-        pass
-
-    def send(self, data: Union[str, bytes, bytearray, int, float, bool]) -> None:
-        pass
-
 
 
 class AirPurifierMode(str, Enum):
